+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.13] - 2026-04-28
+
+### Docs
+- Update README.md
+
+### Other
+- Update regres/__init__.py
+- Update regres/regres_cli.py
+- Update regres/version_check.py
+
 ## [0.1.12] - 2026-04-28
 
 ### Docs